Coded all of backdrop to respective locations without any april tags used.
Ron I don't know if i did this correctly but look in the backdrop code area i have a  ".stopAndAdd(getackdropActions())" and click on that I don't know if it works or not. I also created a different one and its a public class like the lift class but inside of basic auto. The class is not used because i don't know if it would work but tell me which of the two, the getBackdropActions or the Backdrop class will would work. Otherwise i can just copy and paste the backdrop arm stuff in there without doing the extra stuff.